Title: 실수 받아 그대로 출력, Time: -1ms, Memory: -1MB, Status: Runtime Error - CodeTree

+++ 241021/실수 받아 그대로 출력/enter-real-value-and-print.java
@@ -3,7 +3,7 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        int n = sc.nextInt();
+        double n = sc.nextInt();
 
         System.out.printf("%.2f",n);
     }
